@@ -1,69 +1,71 @@
 {
-  "name": "tsukibot",
-  "version": "0.2.3",
-  "description": "Discord bot with cryptocurrency functionalities",
+  "name": "edgebot",
+  "version": "0.1.0",
+  "description": "",
   "main": "src/bot.js",
   "author": "Oscar Fonseca",
-  "repository": "https://github.com/OFRBG/TsukiBot",
+  "repository": "https://github.com/OFRBG/EdgeBot",
   "license": "MIT",
   "engines": {
-    "node": ">=10.0.0"
+    "node": ">=14.0.0"
   },
   "scripts": {
-    "start": "node ./src/bot.js",
-    "dev": "nodemon -- -d",
-    "lint": "eslint src",
-    "flow": "flow",
-    "deploy": "gcloud app deploy app.base.yaml --version staging --quiet",
-    "version": "HUSKY_SKIP_HOOKS=1 standard-version"
+    "start": "ts-node -r dotenv-flow/config -r tsconfig-paths/register src/bot.ts",
+    "dev": "nodemon --watch 'src/**/*.ts' --ignore 'src/**/*.spec.ts' --exec 'ts-node -r dotenv-flow/config -r tsconfig-paths/register' src/bot.ts -- -d",
+    "debug": "nodemon --watch 'src/**/*.ts' --ignore 'src/**/*.spec.ts' --exec 'node --inspect -r dotenv-flow/config -r ts-node/register -r tsconfig-paths/register' src/bot.ts",
+    "build": "run-s build:*",
+    "build:out": "ttsc -p tsconfig.json",
+    "fix": "run-s fix:*",
+    "fix:prettier": "prettier \"src/**/*.ts\" --write",
+    "fix:lint": "eslint --fix --ext .ts ./src",
+    "test": "run-s build test:*",
+    "test:lint": "eslint --ext .ts ./src && prettier \"src/**/*.ts\" --list-different",
+    "test:unit": "jest",
+    "version": "HUSKY_SKIP_HOOKS=1 standard-version",
+    "madge": "madge src/bot.ts --exclude 'globals/index.ts|logger.ts' --ts-config ./tsconfig.json --dot | dot -Nshape=box -Gsplines=ortho -Gnewrank -Granksep=1 -Gnodesep=0.7 -Nwidth=2 -Tpng >| deps.png"
   },
   "dependencies": {
-    "@google-cloud/logging-winston": "^3.0.2",
-    "bitfinex-api-node": "^1.2.1",
-    "coinbase": "^2.0.8",
-    "coinmarketcap": "^0.2.0",
-    "cryptocompare": "^0.1.0",
-    "didyoumean": "^1.2.1",
-    "discord.js": "^11.3.2",
-    "dotenv-flow": "^3.1.0",
-    "env-var": "^5.1.0",
-    "gdax": "^0.5.1",
-    "kraken-api": "^1.0.0",
-    "lodash": "^4.17.15",
-    "node-binance-api": "^0.3.9",
-    "node-fetch": "^1.7.3",
-    "node-schedule": "^1.3.0",
-    "node.bittrex.api": "^0.5.3",
-    "request": "^2.85.0",
-    "request-promise-native": "^1.0.8",
-    "winston": "^3.2.1",
-    "ws": "^5.1.1",
-    "yargs": "^15.0.1"
+    "@types/axios": "^0.14.0",
+    "@types/dotenv-flow": "^3.0.0",
+    "@types/lodash": "^4.14.159",
+    "@types/winston": "^2.4.4",
+    "@types/yargs": "^15.0.5",
+    "axios": "^0.19.2",
+    "discord.js": "^12.2.0",
+    "dotenv-flow": "^3.2.0",
+    "env-var": "^6.3.0",
+    "lodash": "^4.17.19",
+    "npm-run-all": "^4.1.5",
+    "ts-node": "^8.10.2",
+    "tsconfig-paths": "^3.9.0",
+    "winston": "^3.3.3",
+    "yargs": "^15.4.1"
   },
   "devDependencies": {
-    "commitizen": "^4.0.3",
+    "@types/jest": "^26.0.7",
+    "@types/node": "^14.0.26",
+    "@typescript-eslint/eslint-plugin": "^3.7.0",
+    "@typescript-eslint/parser": "^3.7.0",
+    "commitizen": "^4.1.2",
     "cz-conventional-changelog": "^3.0.2",
-    "eslint": "^6.5.1",
-    "eslint-config-airbnb": "^18.0.1",
-    "eslint-config-airbnb-base": "^14.0.0",
+    "eslint": "^7.6.0",
     "eslint-config-prettier": "^6.4.0",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
+    "eslint-plugin-functional": "^3.0.1",
+    "eslint-plugin-jest": "^23.18.2",
     "eslint-plugin-prettier": "^3.1.1",
-    "eslint-plugin-react": "^7.16.0",
-    "eslint-plugin-react-hooks": "^1.7.0",
-    "flow-bin": "^0.114.0",
-    "flow-remove-types": "^2.112.0",
     "husky": "^3.1.0",
-    "jest": "^24.9.0",
-    "nodemon": "^2.0.2",
+    "jest": "^26.1.0",
+    "nodemon": "^2.0.4",
     "prettier": "^1.18.2",
-    "standard-version": "^8.0.1"
+    "source-map-support": "^0.5.19",
+    "standard-version": "^8.0.1",
+    "ts-jest": "^26.1.3",
+    "typescript": "^3.9.7"
   },
   "husky": {
     "hooks": {
       "prepare-commit-msg": "exec < /dev/tty && git cz --hook || true",
-      "pre-commit": "npm run lint"
+      "pre-commit": "npm run fix"
     }
   },
   "config": {
