@@ -1,22 +1,28 @@
 module.exports = {
   env: {
-    node: true,
-    commonjs: true,
-    es6: true,
+    node: true
   },
-  plugins: ['prettier'],
-  extends: [
-    'airbnb-base',
-    'plugin:prettier/recommended'
-  ],
-  globals: {
-    Atomics: 'readonly',
-    SharedArrayBuffer: 'readonly',
-  },
+  parser: "@typescript-eslint/parser",
   parserOptions: {
-    ecmaVersion: 2018,
+    project: "./tsconfig.json",
+    sourceType: "module"
   },
+  plugins: ["@typescript-eslint", "prettier"],
+  extends: [
+    "eslint:recommended",
+    "plugin:prettier/recommended",
+    "plugin:@typescript-eslint/eslint-recommended",
+    "plugin:@typescript-eslint/recommended",
+    "plugin:@typescript-eslint/recommended-requiring-type-checking"
+  ],
   rules: {
-    'prettier/prettier': ['error', { 'singleQuote': true }],
+    "@typescript-eslint/no-misused-promises": [
+      "error",
+      { checksVoidReturn: false }
+    ]
+  },
+  globals: {
+    Atomics: "readonly",
+    SharedArrayBuffer: "readonly"
   }
 };
